+++ p2panda-rs/src/schema/schema.rs
@@ -204,11 +204,12 @@ mod tests {
         // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
         let fields = PinnedRelationList::new(vec![
-            DocumentViewId::new(&[relation_operation_id_1.clone()]),
+            DocumentViewId::new(&[relation_operation_id_1.clone()]).unwrap(),
             DocumentViewId::new(&[
                 relation_operation_id_2.clone(),
                 relation_operation_id_3.clone(),
-            ]),
+            ])
+            .unwrap(),
         ]);
 
         let schema_view = create_schema_view(&fields, &schema_view_id, &field_operation_id);
@@ -229,7 +230,7 @@ mod tests {
         let capacity_field_view = create_field(
             "capacity",
             "int",
-            &DocumentViewId::new(&[relation_operation_id_2, relation_operation_id_3]),
+            &DocumentViewId::new(&[relation_operation_id_2, relation_operation_id_3]).unwrap(),
             &field_operation_id,
         );