+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Highlights are marked with a pancake 🥞
 - Implement schema hash id as a unique identifier for schemas `rs` [#282](https://github.com/p2panda/p2panda/pull/282) `rs`
 - `Graph` method for selecting sub-section of graph [#335](https://github.com/p2panda/p2panda/pull/335) `rs`
 - Storage traits for documents [#343](https://github.com/p2panda/p2panda/pull/343) `rs`
+- Materialise a document at a specific document view [#337](https://github.com/p2panda/p2panda/pull/337) `rs`
 
 ## Changed
 
@@ -58,6 +59,7 @@ Highlights are marked with a pancake 🥞
 - Additional constructor method for `OperationWithMeta` [322](https://github.com/p2panda/p2panda/pull/322) `rs`
 - Minor method renaming in `EntryStore` [323](https://github.com/p2panda/p2panda/pull/323) `rs`
 - Require storage provider errors to be thread-safe [#340](https://github.com/p2panda/p2panda/pull/340)
+- Make previous_operations a `DocumentViewId` [#342](https://github.com/p2panda/p2panda/pull/342) `rs`
 
 ## Fixed